+++ src/FrenetOptimalTrajectory/FrenetOptimalTrajectory.cpp
@@ -53,7 +53,7 @@ FrenetOptimalTrajectory::FrenetOptimalTrajectory(
     }
 
     // select the best path
-    double mincost = INFINITY;
+    float mincost = INFINITY;
     for (FrenetPath *fp : frenet_paths) {
         if (fp->cf <= mincost) {
             mincost = fp->cf;
@@ -61,7 +61,7 @@ FrenetOptimalTrajectory::FrenetOptimalTrajectory(
         }
     }
     auto end = chrono::high_resolution_clock::now();
-    double run_time =
+    float run_time =
         chrono::duration_cast<chrono::nanoseconds>(end - start).count();
     run_time *= 1e-6;
     // cout << "Planning runtime " << run_time << "\n";
@@ -127,17 +127,17 @@ void FrenetOptimalTrajectory::threaded_calc_all_frenet_paths() {
 void FrenetOptimalTrajectory::calc_frenet_paths(int start_di_index,
                                                 int end_di_index,
                                                 bool multithreaded) {
-    double t, ti, tv;
-    double lateral_deviation, lateral_velocity, lateral_acceleration,
+    float t, ti, tv;
+    float lateral_deviation, lateral_velocity, lateral_acceleration,
         lateral_jerk;
-    double longitudinal_acceleration, longitudinal_jerk;
+    float longitudinal_acceleration, longitudinal_jerk;
     FrenetPath *fp, *tfp;
     int num_paths = 0;
     int num_viable_paths = 0;
-    // double valid_path_time = 0;
+    // float valid_path_time = 0;
 
     // initialize di, with start_di_index
-    double di = -fot_hp->max_road_width_l + start_di_index * fot_hp->d_road_w;
+    float di = -fot_hp->max_road_width_l + start_di_index * fot_hp->d_road_w;
 
     // generate path to each offset goal
     // note di goes up to but not including end_di_index*fot_hp->d_road_w
@@ -170,11 +170,11 @@ void FrenetOptimalTrajectory::calc_frenet_paths(int start_di_index,
                 lateral_jerk += abs(lat_qp.calc_third_derivative(t));
                 t += fot_hp->dt;
                 #ifdef USE_RECORDER
-                    Recorder::getInstance()->saveData<double>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::t", fp->t.back());
-                    Recorder::getInstance()->saveData<double>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::d", fp->d.back());
-                    Recorder::getInstance()->saveData<double>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::d_d", fp->d_d.back());
-                    Recorder::getInstance()->saveData<double>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::d_dd", fp->d_dd.back());
-                    Recorder::getInstance()->saveData<double>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::d_ddd", fp->d_ddd.back());
+                    Recorder::getInstance()->saveData<float>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::t", fp->t.back());
+                    Recorder::getInstance()->saveData<float>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::d", fp->d.back());
+                    Recorder::getInstance()->saveData<float>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::d_d", fp->d_d.back());
+                    Recorder::getInstance()->saveData<float>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::d_dd", fp->d_dd.back());
+                    Recorder::getInstance()->saveData<float>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::d_ddd", fp->d_ddd.back());
                 #endif
             }
 
@@ -196,7 +196,7 @@ void FrenetOptimalTrajectory::calc_frenet_paths(int start_di_index,
                     fot_ic->s0, fot_ic->c_speed, 0.0, tv, 0.0, ti);
 
                 // longitudinal motion
-                for (double tp : tfp->t) {
+                for (float tp : tfp->t) {
                     tfp->s.push_back(lon_qp.calc_point(tp));
                     tfp->s_d.push_back(lon_qp.calc_first_derivative(tp));
                     tfp->s_dd.push_back(lon_qp.calc_second_derivative(tp));
@@ -206,10 +206,10 @@ void FrenetOptimalTrajectory::calc_frenet_paths(int start_di_index,
                     longitudinal_jerk += abs(lon_qp.calc_third_derivative(tp));
 
                     #ifdef USE_RECORDER
-                        Recorder::getInstance()->saveData<double>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::s", tfp->s.back());
-                        Recorder::getInstance()->saveData<double>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::s_d", tfp->s_d.back());
-                        Recorder::getInstance()->saveData<double>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::s_dd", tfp->s_dd.back());
-                        Recorder::getInstance()->saveData<double>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::s_ddd", tfp->s_ddd.back());
+                        Recorder::getInstance()->saveData<float>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::s", tfp->s.back());
+                        Recorder::getInstance()->saveData<float>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::s_d", tfp->s_d.back());
+                        Recorder::getInstance()->saveData<float>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::s_dd", tfp->s_dd.back());
+                        Recorder::getInstance()->saveData<float>("FrenetOptimalTrajectory::calc_frenet_paths()::FrenetPath::s_ddd", tfp->s_ddd.back());
                     #endif	
                 }
 
@@ -295,10 +295,10 @@ void FrenetOptimalTrajectory::calc_frenet_paths(int start_di_index,
 
 void FrenetOptimalTrajectory::setObstacles() {
     // Construct obstacles
-    vector<double> llx(fot_ic->o_llx, fot_ic->o_llx + fot_ic->no);
-    vector<double> lly(fot_ic->o_lly, fot_ic->o_lly + fot_ic->no);
-    vector<double> urx(fot_ic->o_urx, fot_ic->o_urx + fot_ic->no);
-    vector<double> ury(fot_ic->o_ury, fot_ic->o_ury + fot_ic->no);
+    vector<float> llx(fot_ic->o_llx, fot_ic->o_llx + fot_ic->no);
+    vector<float> lly(fot_ic->o_lly, fot_ic->o_lly + fot_ic->no);
+    vector<float> urx(fot_ic->o_urx, fot_ic->o_urx + fot_ic->no);
+    vector<float> ury(fot_ic->o_ury, fot_ic->o_ury + fot_ic->no);
 
     for (int i = 0; i < fot_ic->no; i++) {
         addObstacle(Vector2f(llx[i], lly[i]), Vector2f(urx[i], ury[i]));
