+++ src/Polynomials/QuarticPolynomial.cpp
@@ -8,8 +8,8 @@
 
 using namespace Eigen;
 
-QuarticPolynomial::QuarticPolynomial(double xs, double vxs, double axs,
-        double vxe, double axe, double t):
+QuarticPolynomial::QuarticPolynomial(float xs, float vxs, float axs,
+        float vxe, float axe, float t):
         a0(xs), a1(vxs) {
     a2 = axs / 2.0;
     Matrix2d A;
@@ -22,26 +22,26 @@ QuarticPolynomial::QuarticPolynomial(double xs, double vxs, double axs,
     a4 = x[1];
 
     #ifdef USE_RECORDER
-        Recorder::getInstance()->saveData<double>("QuarticPolynomial::QuarticPolynomial::a0", xs);
-        Recorder::getInstance()->saveData<double>("QuarticPolynomial::QuarticPolynomial::a1", vxs);
-        Recorder::getInstance()->saveData<double>("QuarticPolynomial::QuarticPolynomial::a2", a2);
-        Recorder::getInstance()->saveData<double>("QuarticPolynomial::QuarticPolynomial::a3", a3);
-        Recorder::getInstance()->saveData<double>("QuarticPolynomial::QuarticPolynomial::a4", a4);
+        Recorder::getInstance()->saveData<float>("QuarticPolynomial::QuarticPolynomial::a0", xs);
+        Recorder::getInstance()->saveData<float>("QuarticPolynomial::QuarticPolynomial::a1", vxs);
+        Recorder::getInstance()->saveData<float>("QuarticPolynomial::QuarticPolynomial::a2", a2);
+        Recorder::getInstance()->saveData<float>("QuarticPolynomial::QuarticPolynomial::a3", a3);
+        Recorder::getInstance()->saveData<float>("QuarticPolynomial::QuarticPolynomial::a4", a4);
     #endif
 }
 
-double QuarticPolynomial::calc_point(double t) {
+float QuarticPolynomial::calc_point(float t) {
     return a0 + a1 * t + a2 * pow(t, 2) + a3 * pow(t, 3) + a4 * pow(t, 4);
 }
 
-double QuarticPolynomial::calc_first_derivative(double t) {
+float QuarticPolynomial::calc_first_derivative(float t) {
     return a1 + 2 * a2 * t + 3 * a3 * pow(t, 2) + 4 * a4 * pow(t, 3);
 }
 
-double QuarticPolynomial::calc_second_derivative(double t) {
+float QuarticPolynomial::calc_second_derivative(float t) {
     return 2 * a2 + 6 * a3 * t + 12 * a4 * pow(t, 2);
 }
 
-double QuarticPolynomial::calc_third_derivative(double t) {
+float QuarticPolynomial::calc_third_derivative(float t) {
     return 6 * a3 + 24 * a4 * t;
 }
