+++ src/Polynomials/QuinticPolynomial.h
@@ -6,14 +6,14 @@
 class QuinticPolynomial {
 public:
     QuinticPolynomial() = default;
-    QuinticPolynomial(double xs, double vxs, double axs, double xe,
-                      double vxe, double axe, double t);
-    double calc_point(double t);
-    double calc_first_derivative(double t);
-    double calc_second_derivative(double t);
-    double calc_third_derivative(double t);
+    QuinticPolynomial(float xs, float vxs, float axs, float xe,
+                      float vxe, float axe, float t);
+    float calc_point(float t);
+    float calc_first_derivative(float t);
+    float calc_second_derivative(float t);
+    float calc_third_derivative(float t);
 private:
-    double a0, a1, a2, a3, a4, a5;
+    float a0, a1, a2, a3, a4, a5;
 };
 
 #endif //FRENET_OPTIMAL_TRAJECTORY_QUINTICPOLYNOMIAL_H
